Upgraded coffeedoc dependency to ~0.3.1.
Regenerated docs as a result.

coffeedoc used the now-renamed path.existsSync, which is now
fs.existsSync in node >= 0.10.0. This is fixed in 0.3.0.

Unfortunately coffeedoc 0.3.x currently has a dependency on the
highlight package, which is no longer maintained (coffeedoc itself
doesn't seem to be maintained). But this is the best that can be done
without further changes. (Ideally a new documentation library.)